// apps/venture/src/plugins/store/ui/generated-types.ts
export type Exact<T extends { [key: string]: unknown }> = {
	[K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
	[SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
	[SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
	ID: string;
	String: string;
	Boolean: boolean;
	Int: number;
	Float: number;
	/** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
	JSON: any;
	/** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
	DateTime: any;
	/** The `Upload` scalar type represents a file upload. */
	Upload: any;
};
// ...
export type Node = {
	id: Scalars['ID'];
};
// ...
export type Store = Node & {
	__typename?: 'Store';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	name: Scalars['String'];
};

export type StoreList = PaginatedList & {
	__typename?: 'StoreList';
	items: Array<Store>;
	totalItems: Scalars['Int'];
};

export type CreateStoreInput = {
	name: Scalars['String'];
};

export type UpdateStoreInput = {
	id: Scalars['ID'];
	name: Scalars['String'];
};
// ...
export namespace CreateStore {
	export type Variables = CreateStoreMutationVariables;
	export type Mutation = CreateStoreMutation;
	export type CreateStore = NonNullable<CreateStoreMutation['createStore']>;
}

export namespace UpdateStore {
	export type Variables = UpdateStoreMutationVariables;
	export type Mutation = UpdateStoreMutation;
	export type UpdateStore = NonNullable<UpdateStoreMutation['updateStore']>;
}
// ...
export type CreateStoreMutationVariables = Exact<{
	input: CreateStoreInput;
}>;

export type CreateStoreMutation = { __typename?: 'Mutation' } & {
	createStore: { __typename?: 'Store' } & Pick<Store, 'id' | 'name'>;
};

export type UpdateStoreMutationVariables = Exact<{
	input: UpdateStoreInput;
}>;

export type UpdateStoreMutation = { __typename?: 'Mutation' } & {
	updateStore: { __typename?: 'Store' } & Pick<Store, 'id' | 'name'>;
};